+++ src/formatters.rs
@@ -78,18 +78,24 @@ pub fn s2v_compression_ratio() -> Arc<dyn Fn(&str) -> Option<f32> + Send + Sync>
 }
 
 /// Turn an `f32` value from voltage gain to decibels using the semantics described in
-/// [`util::gain_to_db()]. You should use either `" dB"` or `" dBFS"` for the parameter's unit.
-/// `0.0` will be formatted as `-inf`.
+/// [`util::gain_to_db()`]. You should use either `" dB"` or `" dBFS"` for the parameter's unit.
+/// `0.0` will be formatted as `-inf`. Avoids returning negative zero values to make sure
+/// string->value->string roundtrips work correctly. Otherwise `-0.001` rounded to two digits
+/// would result in `-0.00`.
 pub fn v2s_f32_gain_to_db(digits: usize) -> Arc<dyn Fn(f32) -> String + Send + Sync> {
+    let rounding_multiplier = 10u32.pow(digits as u32) as f32;
     Arc::new(move |value| {
         if value < util::MINUS_INFINITY_GAIN {
             String::from("-inf")
         } else {
-            // Never print -0.0 since that just looks weird and confusing
             let value_db = util::gain_to_db(value);
-            let value_db = if value_db.abs() < 1e-6 { 0.0 } else { value_db };
 
-            format!("{value_db:.digits$}")
+            // See above
+            if (value_db * rounding_multiplier).round() / rounding_multiplier == 0.0 {
+                format!("{:.digits$}", 0.0)
+            } else {
+                format!("{value_db:.digits$}")
+            }
         }
     })
 }
@@ -119,7 +125,7 @@ pub fn v2s_f32_panning() -> Arc<dyn Fn(f32) -> String + Send + Sync> {
     })
 }
 
-/// Parse a pan value in the format of [`v2s_f32_panning()] to a linear value in the range `[-1,
+/// Parse a pan value in the format of [`v2s_f32_panning()`] to a linear value in the range `[-1,
 /// 1]`.
 pub fn s2v_f32_panning() -> Arc<dyn Fn(&str) -> Option<f32> + Send + Sync> {
     Arc::new(|string| {
@@ -190,7 +196,7 @@ pub fn v2s_f32_hz_then_khz_with_note_name(
     })
 }
 
-/// Convert an input in the same format at that of [`v2s_f32_hz_then_khz()] to a Hertz value. This
+/// Convert an input in the same format at that of [`v2s_f32_hz_then_khz()`] to a Hertz value. This
 /// additionally also accepts note names in the same format as [`s2v_i32_note_formatter()`], and
 /// optionally also with cents in the form of `D#5, -23 ct.`.
 pub fn s2v_f32_hz_then_khz() -> Arc<dyn Fn(&str) -> Option<f32> + Send + Sync> {
@@ -271,7 +277,7 @@ pub fn v2s_i32_note_formatter() -> Arc<dyn Fn(i32) -> String + Send + Sync> {
     })
 }
 
-/// Parse a note name to a MIDI number using the inverse mapping from [`v2s_i32_note_formatter()].
+/// Parse a note name to a MIDI number using the inverse mapping from [`v2s_i32_note_formatter()`].
 pub fn s2v_i32_note_formatter() -> Arc<dyn Fn(&str) -> Option<i32> + Send + Sync> {
     Arc::new(|string| {
         let string = string.trim();
@@ -316,7 +322,7 @@ pub fn v2s_bool_bypass() -> Arc<dyn Fn(bool) -> String + Send + Sync> {
     })
 }
 
-/// Parse a string in the same format as [`v2s_bool_bypass()].
+/// Parse a string in the same format as [`v2s_bool_bypass()`].
 pub fn s2v_bool_bypass() -> Arc<dyn Fn(&str) -> Option<bool> + Send + Sync> {
     Arc::new(|string| {
         let string = string.trim();